Added login and register functionality

+++ base/views.py
@@ -1,6 +1,11 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.http import HttpResponse
 from .models import Room, Topic
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 from .forms import RoomForm
 
 # rooms = [
@@ -8,6 +13,42 @@ from .forms import RoomForm
 #     {'id' : 2, 'name': 'Design with me!'},
 #     {'id' : 3, 'name': 'Frontend development'},
 # ]
+
+def loginPage(request):
+
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password') #these two values will be sent from the frontend
+
+        try: #to make sure whether the user exists or not we use try except block
+            user = User.objects.get(username = username)
+        except:
+            messages.error(request, 'User does not exist')
+
+            user = authenticate(request, username = username, password = password)
+
+            if user is not None:
+                login(request, user)
+                return redirect('home')
+            else:
+                messages.error(request, 'Username or password does not exist')
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    page = 'register'
+    return render(request, 'base/login_register.html')
 
 
 def home(request):
@@ -29,6 +70,8 @@ def room(request, pk):
     room = None         #room is set to None because at first there will be no rooms
     return render(request, 'base/room.html')
 
+
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -41,9 +84,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk) #update the room by id
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance = room)
@@ -54,8 +102,14 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
